worked on notification system

+++ src/Controller/ApiController.php
@@ -6,6 +6,7 @@ use App\Entity\Group;
 use App\Entity\Notification;
 use App\Entity\User;
 use App\Repository\GroupRepository;
+use App\Repository\NotificationRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use App\Repository\UserRepository;
 use Doctrine\ORM\EntityManagerInterface;
@@ -17,10 +18,9 @@ use Symfony\Component\HttpFoundation\Request;
 #[Route('/api', name: 'app_api')]
 class ApiController extends AbstractController
 {
-    #[Route('/searchUpdater/{searchedUser?}', name: 'app_searchUpdater', methods: ['GET'])]
-    public function searchUpdater($searchedUser, UserRepository $userRepo): JsonResponse
+    #[Route('/user-search/{searchedUser?}', name: 'app_userSearch', methods: ['GET'])]
+    public function userSearch($searchedUser, UserRepository $userRepo): JsonResponse
     {
-        $this->denyAccessUnlessGranted('ROLE_USER');
         $result = $userRepo->findBySearchBar($searchedUser);
         $cleanedResult = [];
         foreach($result as $user){
@@ -29,6 +29,23 @@ class ApiController extends AbstractController
                 'displayedName' => $user->getDisplayedName(),
                 'avatar' => $user->getAvatar(),
             );
+        }
+        return $this->json(json_encode($cleanedResult));
+    }
+
+    #[Route('/group-search', name: 'app_groupSearch', methods: ['GET'])]
+    public function groupSearch(): JsonResponse
+    {
+        /**
+         * @var User
+         */
+        $user = $this->getUser();
+        $result = $user->getGroups();
+        $cleanedResult = [];
+        foreach($result as $group){
+            if($group->getOwner() == $user){
+                $cleanedResult[] = $group->getName();
+            }
         }
         return $this->json(json_encode($cleanedResult));
     }
@@ -86,17 +103,17 @@ class ApiController extends AbstractController
 
         // Return error if user try to add himself as contact
         if($user == $contact){
-            return new JsonResponse(['error' => 'you can\'t remove yourself in contact list'], 403);
+            return new JsonResponse(['error' => 'You can\'t remove yourself in contact list'], 403);
         }
 
         // Return error if user try to add a contact that doesn't exist
         if(!$contact){
-            return new JsonResponse(['error' => 'contact doesn\'t exist'], 404);
+            return new JsonResponse(['error' => 'Contact does not exist'], 404);
         }
 
         // Return error if user try to add a contact already in contact list
         if(!$user->getContact()->contains($contact)){
-            return new JsonResponse(['error' => 'contact already not in contact list'], 403);
+            return new JsonResponse(['error' => 'Contact is already in contact list'], 403);
         }
 
         // Add contact and sent to db
@@ -107,70 +124,73 @@ class ApiController extends AbstractController
         return new JsonResponse(['status' => 'Contact removed'], 200);
     }
 
-    #[Route('/change-group-name/{groupId}', name: 'app_changeGroupName', methods: ['POST'])]
-    public function changeGroupName($groupId, Request $request, GroupRepository $groupRepo, EntityManagerInterface $em): JsonResponse
+    #[Route('/change-group-name', name: 'app_changeGroupName', methods: ['PATCH'])]
+    public function changeGroupName(Request $request, GroupRepository $groupRepo, EntityManagerInterface $em): JsonResponse
     {
-        $group = $groupRepo->find($groupId);
-        if(!$group){
-            return new JsonResponse(['error' => 'group doesn\'t exist'], 404);
-        }
-
-        if($this->getUser() != $group->getOwner()){
-            return new JsonResponse(['error' => 'your are not the owner'], 403);
-        }
         $data = json_decode($request->getContent(), true);
+        $group = $groupRepo->find($data['groupId']);
+
+        if(!$group){
+            return new JsonResponse(['error' => 'Group does not exist'], 404);
+        }
+        if($this->getUser() != $group->getOwner()){
+            return new JsonResponse(['error' => 'You are not the owner'], 403);
+        }
+        
         $group->setName($data['name']);
         $em->persist($group);
         $em->flush();
     }
 
-    #[Route('/change-description/{groupId}', name: 'app_changeDescription', methods: ['POST'])]
-    public function changeDescription($groupId, Request $request, GroupRepository $groupRepo, EntityManagerInterface $em): JsonResponse
+    #[Route('/change-description', name: 'app_changeDescription', methods: ['PATCH'])]
+    public function changeDescription(Request $request, GroupRepository $groupRepo, EntityManagerInterface $em): JsonResponse
     {
+        $data = json_decode($request->getContent(), true);
+        $groupId = $data['groupId'];
         $group = $groupRepo->find($groupId);
+
         if(!$group){
-            return new JsonResponse(['error' => 'group doesn\'t exist'], 404);
+            return new JsonResponse(['error' => 'Group does not exist'], 404);
+        }
+        if($this->getUser() != $group->getOwner()){
+            return new JsonResponse(['error' => 'You are not the owner'], 403);
         }
 
-        if($this->getUser() != $group->getOwner()){
-            return new JsonResponse(['error' => 'your are not the owner'], 403);
-        }
-        $data = json_decode($request->getContent(), true);
         $group->setDescription($data['description']);
         $em->persist($group);
         $em->flush();
     }
 
-    #[Route('/delete-group/{groupId}', name: 'app_deleteGroup', methods: ['POST'])]
-    public function deleteGroup($groupId, Request $request, GroupRepository $groupRepo, EntityManagerInterface $em): JsonResponse
+    #[Route('/delete-group', name: 'app_deleteGroup', methods: ['DELETE'])]
+    public function deleteGroup(Request $request, GroupRepository $groupRepo, EntityManagerInterface $em): JsonResponse
     {
+        $groupId = json_decode($request->getContent(), true)['groupId'];
         $group = $groupRepo->find($groupId);
         if(!$group){
-            return new JsonResponse(['error' => 'group doesn\'t exist'], 404);
+            return new JsonResponse(['error' => 'Group does not exist'], 404);
         }
 
         if($this->getUser() != $group->getOwner()){
-            return new JsonResponse(['error' => 'your are not the owner'], 403);
+            return new JsonResponse(['error' => 'You are not the owner'], 403);
         }
-        $data = json_decode($request->getContent(), true);
-        $group->setDescription($data['description']);
-        $em->persist($group);
+        $em->remove($group);
         $em->flush();
     }
     
-    #[Route('invite/{groupId}/{userId}', name: 'app_inviteInGroup', methods: ['POST'])]
-    public function inviteInGroup($groupId, $userId, GroupRepository $groupRepo, UserRepository $userRepo,EntityManagerInterface $em): JsonResponse
-    {
+    #[Route('/invite', name: 'app_inviteInGroup', methods: ['GET'])]
+    public function inviteInGroup(Request $request, GroupRepository $groupRepo, UserRepository $userRepo,EntityManagerInterface $em): JsonResponse
+    {   
+        $data = json_decode($request->getContent(), true);
         // Check if group exist
-        $group = $groupRepo->find($groupId);
+        $group = $groupRepo->find($data['groupId']);
         if(!$group){
-            return new JsonResponse(['error' => 'group doesn\'t exist'], 404);
+            return new JsonResponse(['error' => 'Group doesn\'t exist'], 404);
         }
 
         // Check if invited user exist
-        $userInvited = $userRepo->find($userId);
+        $userInvited = $userRepo->find($data['userId']);
         if(!$userInvited){
-            return new JsonResponse(['error' => 'user doesn\'t exist'], 404);
+            return new JsonResponse(['error' => 'User doesn\'t exist'], 404);
         }
 
         /**
@@ -178,9 +198,9 @@ class ApiController extends AbstractController
         **/
         $user = $this->getUser();
 
-        //Check if user has rights to invite
+        // Check if user has rights to invite
         if($user != $group->getOwner()){
-            return new JsonResponse(['error' => 'your are not the owner'], 403);
+            return new JsonResponse(['error' => 'You are not the owner'], 403);
         }
         
         // Send notification to user to ask him to join
@@ -194,45 +214,64 @@ class ApiController extends AbstractController
         $em->flush();
     }
 
-    #[Route('acceptInvite/{notificationId}', name: 'app_acceptInvite', methods: ['GET'])]
-    public function acceptInvite($notificationId, GroupRepository $groupRepo,EntityManagerInterface $em): JsonResponse
+    #[Route('/accept-invite', name: 'app_acceptInvite', methods: ['POST'])]
+    public function acceptInvite(Request $request, GroupRepository $groupRepo,EntityManagerInterface $em): JsonResponse
     {
         /**
         * @var User
         */
         $user = $this->getUser();
+        $notificationId = json_decode($request->getContent(), true)['notificationId'];
+
 
         // Search if user got an invite
         foreach($user->getNotifications() as $notif){
-            if($notif->getId() == $notificationId){
+            if($notif->getId() == $notificationId && $notif->getType() == 'group'){
                 $group = $groupRepo->find($notif->getIsGroupId());
                 $group->addMember($user);
                 $em->persist($group);
                 $em->remove($notif);
                 $em->flush();
-                return new JsonResponse(['success' => 'invite successfully accepted'], 200);
+                return new JsonResponse(['success' => 'Invite successfully accepted'], 200);
             }
         }
-        return new JsonResponse(['error' => 'invite doesn\'t exist'], 404);
+        return new JsonResponse(['error' => 'Invite does not exist'], 404);
     }
 
-    #[Route('declineInvite/{notificationId}', name: 'app_declineInvite', methods: ['GET'])]
-    public function declineInvite($notificationId, EntityManagerInterface $em): JsonResponse
+    #[Route('/decline-invite', name: 'app_declineInvite', methods: ['POST'])]
+    public function declineInvite(Request $request, EntityManagerInterface $em): JsonResponse
     {
         /**
         * @var User
         */
         $user = $this->getUser();
+        $notificationId = json_decode($request->getContent(), true)['notificationId'];
 
         //search if user got an invite
         foreach($user->getNotifications() as $notif){
-            if($notif->getId() == $notificationId){
+            if($notif->getId() == $notificationId && $notif->getType() == 'group'){
                 
                 $em->remove($notif);
                 $em->flush();
-                return new JsonResponse(['success' => 'invite successfully declined'], 200);
+                return new JsonResponse(['success' => 'Invite successfully declined'], 200);
             }
         }
-        return new JsonResponse(['error' => 'invite doesn\'t exist'], 404);
+        return new JsonResponse(['error' => 'Invite does not exist'], 404);
+    }
+
+    #[Route('/notification-read-change', name: 'app_notificationreadChange', methods: ['PATCH'])]
+    public function NotificationreadChange(Request $request, NotificationRepository $notifRepo, EntityManagerInterface $em): JsonResponse
+    {
+        $data = json_decode($request->getContent(), true);
+        $notificationId = $data['notificationId'];
+        $notification = $notifRepo->find($notificationId);
+
+        if(!$notification){
+            return new JsonResponse(['error' => 'This notification does not exist'], 404);
+        }
+
+        $notification->setRead($data['state']);
+        dd($data['state']);
+        return new JsonResponse(['status' => 'Read state changed'], 200);
     }
 }
